fix: hydration error on countdown
